[show-hint addon] fixed fromList didn't show all completions if the cursor was inside the token
The hint logic of the `fromList` command did not respect the cursor position inside the token.
Now it will list all matching completions for the substring up to the cursor, instead of only the ones matching the complete token. The behavior is now the same as for the regular autocompletion.