[merge addon] Use CSS :after to style scroll-lock icon
Primary motivation was to remove the use of `innerHTML` because the
Firefox extension validator raises a warning about this.

The obvious solution is to use `textContent` instead of `innerHTML`
but I thought it could be useful to have the ability to pick a
custom visual for the scroll-lock icon.

As per MDN documentation, all the supported browsers as per
CodeMirror documentation support `:after`.