[lint addon] Allow appending the tooltip to the wrapper element
The tooltip in the existing way append to the body.

When applied the codemirror context inside a shadow root (custom
element) this calls append the tooltip out of the root components.

Thus, the user has to monkey patch the appendchild method to override
the behavior.

This change would allow the tooltip to render in its own context with
the `selfContain` option.