[tern addon] Allow appending the tooltip to the codemirror hint options container if exists
The current implementation of tern.js is appending the tooltip to the body.

When applied the codemirror context inside a shadow root, tooltips are falling out of the root component.
This change would append the tooltip to the container of codemirror hint options if it has else it will default to document.body