[soy mode] Update indentation rules for Element Composition
* Add support for Soy Element Composition.

Add support for Soy Element Composition. It has the syntax in the form of

<{foo()}></>

This adds support to pass through allowEmptyTag and to support this mode in closetag.

* Disable allowMissingTagName and handle Soy Element Composition directly.

Disable allowMissingTagName and handle Soy Element Composition directly. This also adds a case in closetag.js to handle autocompletes for soy element composition.

Right now, if you were to do something like

<{foo()}>

it would autocomplet with </undefined>. This change makes it autocomplete with </>

* Update indentation rules for Soy Element Composition

Update indentation rules for Soy Element Composition

* Update soy.js

* Update soy.js